LayoutTests/storage/indexeddb/deletedatabase-transaction.html

<!DOCTYPE html>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
<script>
description("Ensure deleteDatabase() can run concurrently with transactions in other databases");
function test()
{
    removeVendorPrefixes();

    evalAndLog("dbname1 = 'deletedatabase-transaction'");
    evalAndLog("dbname2 = 'deletedatabase-transaction2'");
    debug("");
    debug("Prepare database:");

    request = evalAndLog("indexedDB.deleteDatabase(dbname1)");
    request.onsuccess = function() {
        request = evalAndLog("request = indexedDB.open(dbname1)");
        request.onsuccess = function() {
            evalAndLog("db = request.result");
            request = evalAndLog("db.setVersion('1')");
            request.onsuccess = function () {
                trans = request.result;
                evalAndLog("db.createObjectStore('store')");
                trans.oncomplete = startTransaction;
            };
        };
    };
}

function startTransaction() {
    debug("");
    debug("Start a transaction against the first database:");
    evalAndLog("trans = db.transaction('store', 'readonly')");
    evalAndLog("trans.objectStore('store').get(0)");

    debug("");
    debug("Delete a different database:");
    request = evalAndLog("indexedDB.deleteDatabase(dbname2)");
    request.onblocked = unexpectedBlockedCallback;
    request.onerror = unexpectedErrorCallback;
    request.onsuccess = function() {
        testPassed("success event was fired at delete request");
    };

    trans.oncomplete = finishJSTest;
}

test();
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
