safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/deletedatabase-transaction-expected.txt

CONSOLE MESSAGE: line 190: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
Ensure deleteDatabase() can run concurrently with transactions in other databases

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "deletedatabase-transaction.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
